Make package size fields big integers
We had a package go in today that has an installed size > 2GB, which
makes PostgreSQL blow up when trying to create the package entry.
Enlarge these fields to big integers so they can store all conceivable
package size values.

+++ main/models.py
@@ -169,8 +169,9 @@ class Package(models.Model):
     pkgdesc = models.CharField(max_length=255, null=True)
     url = models.CharField(max_length=255, null=True)
     filename = models.CharField(max_length=255)
-    compressed_size = models.PositiveIntegerField(null=True)
-    installed_size = models.PositiveIntegerField(null=True)
+    # TODO: it would be nice to have the >0 check constraint back here
+    compressed_size = models.BigIntegerField(null=True)
+    installed_size = models.BigIntegerField(null=True)
     build_date = models.DateTimeField(null=True)
     last_update = models.DateTimeField(null=True, blank=True)
     files_last_update = models.DateTimeField(null=True, blank=True)
